+++ src/main/scala/grizzled/zip/package.scala
@@ -1,0 +1,6 @@
+package grizzled
+
+/** The `grizzled.zip` package contains classes and functions to make it easier
+  * to operate on zip and jar files.
+  */
+package object zip